Migrated almost entirely into Spark RDD's ekosystem  - Modified both ReviewsDB, BOWManager, Stemmer, Classifier's to accept RDD's as requests parameters  - Instead of passing Array[Int], now BOWManager returns SparseVector (which is required then by Classifiers - so deal)  - At the moment PolishStemmer is not serializable, so stemming has been splited into three steps with data collecting  - Changed type of accepted marks -> Double, since classifiers use it inside.  -  ...also modified train's accuracy into Double

+++ Analyzer/src/main/scala/review/sentiment/analysis/api/HttpServer.scala
@@ -8,9 +8,9 @@ import scala.concurrent.Future
 
 class HttpServer extends HttpApp {
 
-    var analyzeMethod : String => Future[Int] = _
+    var analyzeMethod : String => Future[Double] = _
 
-    def start(analyzeMethod: String => Future[Int]): Unit = {
+    def start(analyzeMethod: String => Future[Double]): Unit = {
         this.analyzeMethod = analyzeMethod
 
         startServer("localhost", 8080)
